refactor(interactions): interfaces to avoid import cycle
