+++ interactions/command.go
@@ -3,12 +3,8 @@ package interactions
 import (
 	"strconv"
 
-	"github.com/nyttikord/gokord"
-	"github.com/nyttikord/gokord/channel"
 	"github.com/nyttikord/gokord/discord"
 	"github.com/nyttikord/gokord/discord/types"
-	"github.com/nyttikord/gokord/guild"
-	"github.com/nyttikord/gokord/user"
 )
 
 // Command represents an application's slash command.
@@ -108,182 +104,4 @@ type GuildCommandPermissions struct {
 	ApplicationID string                `json:"application_id"`
 	GuildID       string                `json:"guild_id"`
 	Permissions   []*CommandPermissions `json:"permissions"`
-}
-
-// CommandInteractionData contains the data of Command Interaction.
-type CommandInteractionData struct {
-	ID          string                          `json:"id"`
-	Name        string                          `json:"name"`
-	CommandType types.ApplicationCommand        `json:"type"`
-	Resolved    *CommandInteractionDataResolved `json:"resolved"`
-
-	// Slash command options
-	Options []*CommandInteractionDataOption `json:"options"`
-	// InviteTarget (user/message) id on which context menu command was called.
-	// The details are stored in Resolved according to command type.
-	TargetID string `json:"target_id"`
-}
-
-// GetOption finds and returns an CommandOption by its name.
-func (d CommandInteractionData) GetOption(name string) (option *CommandInteractionDataOption) {
-	for _, opt := range d.Options {
-		if opt.Name == name {
-			option = opt
-			break
-		}
-	}
-
-	return
-}
-
-// CommandInteractionDataResolved contains resolved data of Command execution.
-type CommandInteractionDataResolved struct {
-	Users map[string]*user.User `json:"users"`
-	// Partial user.Member are missing user.User, Deaf and Mute fields.
-	Members map[string]*user.Member `json:"members"`
-	Roles   map[string]*guild.Role  `json:"roles"`
-	// Partial channel.Channel only have ID, Name, Type and Permissions fields.
-	Channels    map[string]*channel.Channel           `json:"channels"`
-	Messages    map[string]*channel.Message           `json:"messages"`
-	Attachments map[string]*channel.MessageAttachment `json:"attachments"`
-}
-
-// Type returns the type of InteractionData.
-func (CommandInteractionData) Type() types.Interaction {
-	return types.InteractionApplicationCommand
-}
-
-// CommandInteractionDataOption represents an option of a slash Command.
-type CommandInteractionDataOption struct {
-	Name string                         `json:"name"`
-	Type types.ApplicationCommandOption `json:"type"`
-	// Note: Contains the value specified by Type.
-	Value   interface{}                     `json:"value,omitempty"`
-	Options []*CommandInteractionDataOption `json:"options,omitempty"`
-
-	// Note: autocomplete Interaction only.
-	Focused bool `json:"focused,omitempty"`
-}
-
-// GetOption finds and returns an CommandOption by its name.
-func (o CommandInteractionDataOption) GetOption(name string) (option *CommandInteractionDataOption) {
-	for _, opt := range o.Options {
-		if opt.Name == name {
-			option = opt
-			break
-		}
-	}
-
-	return
-}
-
-// IntValue is a utility function for casting CommandOption value to integer
-func (o CommandInteractionDataOption) IntValue() int64 {
-	if o.Type != types.ApplicationCommandOptionInteger {
-		panic("IntValue called on data option of type " + o.Type.String())
-	}
-	return int64(o.Value.(float64))
-}
-
-// UintValue is a utility function for casting CommandOption value to unsigned integer
-func (o CommandInteractionDataOption) UintValue() uint64 {
-	if o.Type != types.ApplicationCommandOptionInteger {
-		panic("UintValue called on data option of type " + o.Type.String())
-	}
-	return uint64(o.Value.(float64))
-}
-
-// FloatValue is a utility function for casting CommandOption value to float
-func (o CommandInteractionDataOption) FloatValue() float64 {
-	if o.Type != types.ApplicationCommandOptionNumber {
-		panic("FloatValue called on data option of type " + o.Type.String())
-	}
-	return o.Value.(float64)
-}
-
-// StringValue is a utility function for casting CommandOption value to string
-func (o CommandInteractionDataOption) StringValue() string {
-	if o.Type != types.ApplicationCommandOptionString {
-		panic("StringValue called on data option of type " + o.Type.String())
-	}
-	return o.Value.(string)
-}
-
-// BoolValue is a utility function for casting CommandOption value to bool
-func (o CommandInteractionDataOption) BoolValue() bool {
-	if o.Type != types.ApplicationCommandOptionBoolean {
-		panic("BoolValue called on data option of type " + o.Type.String())
-	}
-	return o.Value.(bool)
-}
-
-// ChannelValue is a utility function for casting CommandOption value to channel.Channel.
-// s is the gokord.Session, if not nil, function additionally fetches all channel.Channel's data
-func (o CommandInteractionDataOption) ChannelValue(s *gokord.Session) *channel.Channel {
-	if o.Type != types.ApplicationCommandOptionChannel {
-		panic("ChannelValue called on data option of type " + o.Type.String())
-	}
-	chanID := o.Value.(string)
-
-	if s == nil {
-		return &channel.Channel{ID: chanID}
-	}
-
-	ch, err := s.State.Channel(chanID)
-	if err != nil {
-		ch, err = s.Channel(chanID)
-		if err != nil {
-			return &channel.Channel{ID: chanID}
-		}
-	}
-
-	return ch
-}
-
-// RoleValue is a utility function for casting CommandOption value to guild.Role.
-// s is the gokord.Session, if not nil, function additionally fetches all role's data
-func (o CommandInteractionDataOption) RoleValue(s *gokord.Session, gID string) *guild.Role {
-	if o.Type != types.ApplicationCommandOptionRole && o.Type != types.ApplicationCommandOptionMentionable {
-		panic("RoleValue called on data option of type " + o.Type.String())
-	}
-	roleID := o.Value.(string)
-
-	if s == nil || gID == "" {
-		return &guild.Role{ID: roleID}
-	}
-
-	r, err := s.State.Role(gID, roleID)
-	if err != nil {
-		roles, err := s.GuildRoles(gID)
-		if err == nil {
-			for _, r = range roles {
-				if r.ID == roleID {
-					return r
-				}
-			}
-		}
-		return &guild.Role{ID: roleID}
-	}
-
-	return r
-}
-
-// UserValue is a utility function for casting CommandOption value to user.User.
-// s is the gokord.Session, if not nil, function additionally fetches all user.User's data
-func (o CommandInteractionDataOption) UserValue(s *gokord.Session) *user.User {
-	if o.Type != types.ApplicationCommandOptionUser && o.Type != types.ApplicationCommandOptionMentionable {
-		panic("UserValue called on data option of type " + o.Type.String())
-	}
-	userID := o.Value.(string)
-
-	if s == nil {
-		return &user.User{ID: userID}
-	}
-
-	u, err := s.User(userID)
-	if err != nil {
-		return &user.User{ID: userID}
-	}
-
-	return u
 }
